src/app/landing/landing.component.spec.ts
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { RouterTestingModule } from '@angular/router/testing';

import { LandingComponent } from './landing.component';

describe('LandingComponent', () => {
  let component: LandingComponent;
  let fixture: ComponentFixture<LandingComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [ LandingComponent ],
      imports:[HttpClientTestingModule, RouterTestingModule]
    })
    .compileComponents();

    fixture = TestBed.createComponent(LandingComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });
  it(`should render title`, () => {
    fixture.detectChanges();
    const compiled = fixture.nativeElement as HTMLElement;
    expect(compiled.querySelector('.col-2 h4')?.textContent).toContain('ExploreMore');
  });

  it(`should render signUp today Button`, () => {
    fixture.detectChanges();
    const compiled = fixture.nativeElement as HTMLElement;
    expect(compiled.querySelector('.col-2 button')?.textContent).toContain('SignUp today');
  });
  
  it(`should have as title 'list-of-categries`, () => { 
    const fixture = TestBed.createComponent(LandingComponent); 
    const landing = fixture.componentInstance; 
    expect(landing.categories).toEqual([]); 
                                            
  });
  it(`should render contactInfo`, () => {
    fixture.detectChanges();
    const compiled = fixture.nativeElement as HTMLElement;
    expect(compiled.querySelector('.col-4 p')?.textContent).toContain('Contact Us');
  });

 
  


});
